move HostInfo tests to separate class
